build: migrate package manager from bun to pnpm (#40)
## Why

CI has been red on `main` since May 28 (the E2E and Visual matrices)
with `Electron failed to install correctly`. Root cause: under bun,
Electron's `postinstall` (which downloads the prebuilt binary and writes
`path.txt`) never runs, so `require('electron')` throws. The in-job
retry loop couldn't fix it because the re-install also skipped the
binary download. Build/Lint/Test stayed green only because they don't
launch Electron.

Rather than patch bun's lifecycle-script handling, this moves the
project to **pnpm**, which runs the Electron build script (gated
explicitly via `onlyBuiltDependencies`).

## What changed

- **Package manager → pnpm 11.5.1.** `packageManager` updated;
`bun.lock` replaced with `pnpm-lock.yaml`; new `pnpm-workspace.yaml`
holds the `vite` override (pnpm 11 reads it here, not from the
package.json `pnpm` field) and `allowBuilds` for `electron` + `esbuild`.
- **TS scripts run on Node 24's built-in type stripping** instead of
bun. `tests/visual/run.ts` → `run.mts` (ESM: top-level await needs it,
with `createRequire`/`import.meta.dirname` preserving the old CJS
behaviour). The two `update-platforms` scripts → `.mts`, run via `node`.
No new runtime dependency (tsx not needed).
- **Workflows** (`build`, `lint`, `test`, `publish`, `e2e`,
`visual-tray`): `oven-sh/setup-bun` → `pnpm/action-setup` before
`setup-node` (with `cache: pnpm`); `bun install` → `pnpm install
--frozen-lockfile`; `bun run` → `pnpm run`; `bunx playwright` → `pnpm
exec playwright`. The PR-generating jobs run their `.mts` scripts via
`node` (no install needed).
- **Electron bumped 39 → 42** (devDependency `^42.0.0`, peer range
widened to `<43.0.0`).

## Verification (local)

- `pnpm install` runs Electron's postinstall and downloads the binary.
- `pnpm run build`, `pnpm run typecheck`, `pnpm run lint:check`, `pnpm
run test` (62 passing) — all green.
- Full E2E suite (5 tests) passes against **electron 42.3.2**.
- All 7 workflow YAMLs validate; no `bun` references remain.

Note: the `update-platforms` jobs only run on push to `main`, so this
PR's CI exercises the install + matrices; those PR-bot jobs are verified
by the equivalent `node` runs locally.
