+++ .github/workflows/e2e.yml
@@ -47,42 +47,44 @@ jobs:
               uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
               with:
                   persist-credentials: false
+            - name: Setup pnpm
+              uses: pnpm/action-setup@d15e628ca66d93ee5f352c71671a7bc6a97af5c9 # v6.0.8
             - name: Setup Node.js
               uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
               with:
                   node-version-file: .node-version
-            - name: Setup Bun
-              uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
-              with:
-                  bun-version-file: package.json
+                  cache: pnpm
             - name: Install dependencies
               shell: bash
               run: |
+                  # pnpm does not reliably run electron's postinstall via
+                  # allowBuilds, so invoke its installer explicitly to fetch the
+                  # binary (path.txt). Verify with a require() before proceeding.
                   for attempt in 1 2 3; do
-                      if bun install --frozen-lockfile; then
+                      if pnpm install --frozen-lockfile && node node_modules/electron/install.js && node -e "require('electron')" >/dev/null 2>&1; then
                           exit 0
                       fi
-                      echo "::warning::install attempt $attempt failed (often electron postinstall 502); retrying..."
+                      echo "::warning::install attempt $attempt failed (often electron 502); retrying..."
                       sleep $((attempt * 5))
                   done
                   echo "::error::install failed after 3 attempts"
                   exit 1
             - name: Build library
-              run: bun run build
+              run: pnpm run build
             - name: Run E2E (Linux)
               if: runner.os == 'Linux'
               env:
                   E2E_PLATFORM_KEY: ${{ matrix.key }}
                   E2E_PLATFORM_LABEL: ${{ matrix.label }}
                   GITHUB_RUN_URL: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
-              run: xvfb-run --auto-servernum -- bunx playwright test --config tests/e2e/playwright.config.ts
+              run: xvfb-run --auto-servernum -- pnpm exec playwright test --config tests/e2e/playwright.config.ts
             - name: Run E2E (macOS / Windows)
               if: runner.os != 'Linux'
               env:
                   E2E_PLATFORM_KEY: ${{ matrix.key }}
                   E2E_PLATFORM_LABEL: ${{ matrix.label }}
                   GITHUB_RUN_URL: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
-              run: bunx playwright test --config tests/e2e/playwright.config.ts
+              run: pnpm exec playwright test --config tests/e2e/playwright.config.ts
             - name: Upload platform result
               if: always()
               uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
@@ -105,10 +107,10 @@ jobs:
               uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
               with:
                   persist-credentials: false
-            - name: Setup Bun
-              uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
+            - name: Setup Node.js
+              uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
               with:
-                  bun-version-file: package.json
+                  node-version-file: .node-version
             - name: Download all platform results
               uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8.0.1
               with:
@@ -116,7 +118,7 @@ jobs:
                   merge-multiple: true
                   path: test-results/platforms
             - name: Regenerate PLATFORMS block
-              run: bun run tests/e2e/update-platforms.ts
+              run: node tests/e2e/update-platforms.mts
             - name: Open PR if changed
               uses: peter-evans/create-pull-request@271a8d0340265f705b14b6d32b9829c1cb33d45e # v7.0.8
               with:
