+++ .github/workflows/publish.yml
@@ -21,23 +21,22 @@ jobs:
               uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
               with:
                   persist-credentials: false
+            - name: Setup pnpm
+              uses: pnpm/action-setup@d15e628ca66d93ee5f352c71671a7bc6a97af5c9 # v6.0.8
             - name: Setup Node.js
               uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
               with:
                   node-version-file: .node-version
                   registry-url: https://registry.npmjs.org
-            - name: Setup Bun
-              uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
-              with:
-                  bun-version-file: package.json
+                  cache: pnpm
             - name: Install dependencies
-              run: bun install --frozen-lockfile
+              run: pnpm install --frozen-lockfile
             - name: Type check
-              run: bun run typecheck
+              run: pnpm run typecheck
             - name: Run tests
-              run: bun run test
+              run: pnpm run test
             - name: Build library
-              run: bun run build
+              run: pnpm run build
             - name: Publish to npm
               run: npm publish --provenance --access public
               env:
