+++ package.json
@@ -53,12 +53,12 @@
   },
   "scripts": {
     "build": "vp pack",
-    "deploy": "bun run build && standard-version",
+    "deploy": "pnpm build && standard-version",
     "lint:check": "vp check",
     "lint": "vp check --fix",
     "test": "vp test",
     "test:e2e": "playwright test --config tests/e2e/playwright.config.ts",
-    "test:visual": "bun run tests/visual/run.ts",
+    "test:visual": "node tests/visual/run.mts",
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
@@ -66,7 +66,7 @@
     "@types/pngjs": "^6.0.5",
     "@voidzero-dev/vite-plus-core": "latest",
     "@voidzero-dev/vite-plus-test": "latest",
-    "electron": "^39.0.0",
+    "electron": "^42.0.0",
     "pngjs": "^7.0.0",
     "standard-version": "^9.5.0",
     "typescript": "^5.9.3",
@@ -74,10 +74,7 @@
     "vitest": "npm:@voidzero-dev/vite-plus-test@latest"
   },
   "peerDependencies": {
-    "electron": ">=9.0.0 <40.0.0"
+    "electron": ">=9.0.0 <43.0.0"
   },
-  "overrides": {
-    "vite": "npm:@voidzero-dev/vite-plus-core@latest"
-  },
-  "packageManager": "bun@1.3.11"
+  "packageManager": "pnpm@11.5.1"
 }
