+++ pnpm-workspace.yaml
@@ -1,0 +1,15 @@
+# Force vite to resolve to the Vite+ core bundle. pnpm 11 reads overrides from
+# pnpm-workspace.yaml rather than the package.json "pnpm" field.
+overrides:
+  vite: npm:@voidzero-dev/vite-plus-core@latest
+  # Electron's installer hangs while extracting its binary (never writes
+  # path.txt) on Node 24.16+ because of a yauzl bug. Pin yauzl until the
+  # upstream fix ships. https://github.com/electron/electron/issues/51619
+  yauzl: 3.3.1
+
+# Dependencies allowed to run install scripts. electron downloads its prebuilt
+# binary here; esbuild (pulled in by vite-plus) fetches its platform binary.
+# Without these, pnpm skips the scripts and `require('electron')` throws.
+allowBuilds:
+  electron: true
+  esbuild: true
